+++ lib/gpui/native/generated.ex
@@ -22,6 +22,14 @@ defmodule GPUI.Native.Generated do
         :erlang.nif_error(:nif_not_loaded)
       end
 
+      def frame_token(_runtime, _window_id) do
+        :erlang.nif_error(:nif_not_loaded)
+      end
+
+      def await_frame_after(_runtime, _window_id, _generation, _timeout_ms) do
+        :erlang.nif_error(:nif_not_loaded)
+      end
+
       def stop_runtime(_runtime) do
         :erlang.nif_error(:nif_not_loaded)
       end